Removed the usage of CompFlickable. Added textures for rail switch direction
Previously the Rail Switch used CompFlickable to do it's stuff. That comp handled the UI gizmo, flicking, telling pawns to go flick the switch, and more.
Now it's handled by a WorkGiver.

The only noticeable change in gameplay is that a switch waiting to be flicked will not have the usual icon above it to say as much... because I couldn't figure out how to use Designations like I should.

ex. Map.designationManager.DesignationOn(this, DesignationDefOf.FlickRailSwitch)

I really should learn how to use comps and make a custom comp for the rail switch which can use its own designators or whatever.

Reverted  changelog