Fixed both kinds of rails able to coexist in the same space. Fixed dumping on new Powered Rails.
I just used a placeworker to prevent it from happening.

Ideally they should be able to upgrade the rails in-place.
I think there's a mod which does this for walls: I could look into how to do it in the future.

Added code so rail dumps will ignore powered rails as well as normal ones.