Reworked dump-all code loop
Before it would increase the dump radius when there were still valid cells in which to place items. I flipped the order of for loops to remedy this.

Before:
for each item in container
for each cell in cellsToTry
try dump

After:
for each cell in cellsToTry
for each item in container
try dump

I can't put to words why this fixed the problem, but it did.